Added the double feature in slope. Significant increase in tolerance not seen.

+++ convex_hull_utility.h
@@ -13,7 +13,7 @@ Describing the data structure to hold the slope of pair of points
 struct p2p_slope{
     int idx1;
     int idx2;
-    float slope;
+    double slope;
 };
 
 /*
